# Conflicts:
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/optitem/OptionalItem.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/test/java/nts/uk/ctx/at/shared/dom/scherec/optitem/OptionalItemTest.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/find/monthly/root/dto/WorkAmountOfMonthlyDto.java
@@ -19,7 +19,7 @@ import nts.uk.ctx.at.shared.dom.scherec.monthlyattdcal.monthly.verticaltotal.wor
 public class WorkAmountOfMonthlyDto implements ItemConst, AttendanceItemDataGate {
 	
 	/** 就業時間金額 */
-	private int workTimeAmount;
+	private long workTimeAmount;
 	
 	public static WorkAmountOfMonthlyDto from(WorkAmountOfMonthly domain) {
 		WorkAmountOfMonthlyDto dto = new WorkAmountOfMonthlyDto();
